Remove `AutoBeOpenApi.IOperation.{summary|specification}`
